# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CoopData API",
    "description": "Cooperative Data Management Platform API with RBAC hierarchy (Ministry \u2192 Federation \u2192 Apex \u2192 Cooperative)",
    "contact": {
      "name": "CoopData Team",
      "email": "support@coopdata.org"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/cooperative/financial-statement/upload": {
      "post": {
        "tags": [
          "Cooperative"
        ],
        "operationId": "upload_financial_statement",
        "responses": {
          "202": {
            "description": "Upload accepted, extraction queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or unsupported file type"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Cooperative profile not found"
          },
          "409": {
            "description": "Submission already exists for this year"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UploadResponse": {
        "type": "object",
        "required": [
          "submission_id",
          "financial_statement_id",
          "extraction_job_id"
        ],
        "properties": {
          "extraction_job_id": {
            "type": "string",
            "format": "uuid"
          },
          "financial_statement_id": {
            "type": "string",
            "format": "uuid"
          },
          "submission_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  }
}